+++ django_admin_rest_api/api/views/detail.py
@@ -36,6 +36,7 @@ from django.http import HttpResponse
 from django.http import JsonResponse
 from django.views.generic import View
 
+from django_admin_rest_api.api.actions_meta import actions_payload
 from django_admin_rest_api.api.inlines import inlines_payload
 from django_admin_rest_api.api.permissions import forbidden_response
 from django_admin_rest_api.api.permissions import is_admin_user
@@ -51,7 +52,6 @@ from django_admin_rest_api.api.serializers import label_for
 from django_admin_rest_api.api.serializers import safe_get_field
 from django_admin_rest_api.api.serializers import serialize_fk_value
 from django_admin_rest_api.api.serializers import serialize_value
-from django_admin_rest_api.api.views.actions import actions_payload
 from django_admin_rest_api.api.writes import load_object_or_none
 from django_admin_rest_api.api.writes import not_found_response